further work on Authentication  - add_key  - remove_key  - find_key  - handle_update
to avoid convoluting primary / secondary code with this stuff.  Ensure SOA is
present, and assumes that for each key name there is only a singleton key.

The secondary behaves slightly different now when a key is removed - it does no
longer touch the data store.  It still requests SOA (and modifies its zone map)
upon addition and removal of transfer keys.