ocertify: support the pem file in use to contain a certificate chain
Since ACME v2 the full chain is available via ACME. ocertify.exe stores the
entire chain and leaf certificate in the pem file. Now, reading this pem file
lead to an error since only the leaf certificate was expected.

This commit reads a list of certificates from the file, and picks the one that
supports the hostname for further processing (validity checking, etc.).