dns-client (mirage/lwt): avoid connecting to resolver race
When multiple DNS requests are issued, each of them attempts a connection to
the resolver (since the connection initialization is not tracked), leading to
multiple file descriptors being used and the fd/flow being set to one of the
connections. Instead, once a connection is attempted, establish a Lwt_condition
and wait in other promises until that is resolved before proceeding.