dns_trie: fix lookup when delegations are present, fix entries for delegations
if a.example is delegated (a.example NS foo), and a record (glue) exists
(a.example A 1.2.3.4), the Dns_trie.lookup returned it instead of a `Delegation
error. This brings lookup and lookup_with_cname (which uses lookup_res) more
into the same shape (lookup_aux >>= check_zone >>= lookup_in_map)

Dns_trie.entries checked for SOA records, but zone cuts are at the presence of
a NS record -- SOA is only present in the child zone (and thus only present if
the very same NS is responsible for both parent and child).