resolver: Use Blocked EDE over Filtered
Initially, Filtered was chosen because it communicated it was filtered
due to the client requesting that. That sort of made sense in my head
because I was running my own resolver. However, there's no guarantee of
that - it may be a guest on my network etc etc.

Furthermore, pi-hole uses Blocked and the description of Blocked is more
neutral.